README: use go install for newer go versions
This PR updates the README installation section, to use `go install` for newer Go versions (>= 1.17). Both the old and new method remain described to support users of the older Go versions.

As of Go 1.17, the use of `go get` for installing executable is [deprecated]( https://golang.org/doc/go-get-install-deprecation). And will be disabled from Go 1.18. As per message on command line:

```
go get: installing executables with 'go get' in module mode is deprecated.
        Use 'go install pkg@version' instead.
        For more information, see https://golang.org/doc/go-get-install-deprecation
        or run 'go help get' or 'go help install'.
```